pyproject.toml
[project]
name = "pinboard-mcp"
version = "0.1.0"
description = "MCP server for access to Pinboard.in bookmarks"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=0.9.0",
    "pinboard>=2.1.1",
    "python-dateutil>=2.8.0",
    "pydantic>=2.0.0",
    "httpx>=0.25.0"
]

[project.scripts]
pinboard-mcp = "pinboard_mcp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
